Motor indexing
Motor index is now associated in a tuple with the corresponding
FeatureProvider, since motors can have different indices on different
levels. E.g. Motor A has index 1 on IC1 and index 10 on the module.
While we have to take care about motor differentiation on IC-level, the
motor differentiation on module-level is done in firmware.